# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shop GazTochka API",
        "description": "Документация API для интернет-магазина ГазТочка",
        "contact": {
            "email": "admin@gaztochka.ru"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Основной сервер API"
        }
    ],
    "paths": {
...
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Получить список всех категорий",
                "operationId": "4aed8169864e0b796b0ca9e85d852c15",
                "responses": {
                    "200": {
                        "description": "Список категорий",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CategoryResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Создать новую категорию",
                "operationId": "495d1e25b36375b432340ee0c76381e1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Категория успешно создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
...
        "/api/v1/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Получить информацию о категории по ID",
                "operationId": "be0aa803b4ba569796f0d43b6a77b08a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о категории",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Категория не найдена"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Обновить существующую категорию",
                "operationId": "f08e5c40d125d3c884b755921a74d048",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Категория успешно обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Категория не найдена"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Удалить категорию",
                "operationId": "5d53cc6e02e3d6977b148e505f7b7fb4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Категория успешно удалена"
                    },
                    "404": {
                        "description": "Категория не найдена"
                    }
                }
            }
        },
        "/api/v1/categories/{id}/breadcrumbs": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Получить хлебные крошки для категории",
                "operationId": "0eb308641c10287565eb33ffaf41a7a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Хлебные крошки",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CategoryResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Категория не найдена"
                    }
                }
            }
        },
        "/api/v1/categories/{id}/products": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Получить товары категории (включая подкатегории)",
                "operationId": "6f4a8e88fadf2eb10abe916f0c80c7b4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "include_children",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список товаров",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Категория не найдена"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CategoryDTO": {
                "required": [
                    "name",
                    "slug"
                ],
                "properties": {
                    "parent_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "example": "Газовое оборудование"
                    },
                    "slug": {
                        "type": "string",
                        "example": "gazovoe-oborudovanie"
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "description": {
                        "type": "string",
                        "example": "Описание категории",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "CategoryResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "example": "Газовое оборудование"
                    },
                    "slug": {
                        "type": "string",
                        "example": "gazovoe-oborudovanie"
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "description": {
                        "type": "string",
                        "example": "Описание категории",
                        "nullable": true
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CategoryResource"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ProductResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Газовый котел"
                    },
                    "slug": {
                        "type": "string",
                        "example": "gazoviy-kotel"
                    },
                    "description": {
                        "type": "string",
                        "example": "Описание товара",
                        "nullable": true
                    },
                    "attributes": {
                        "type": "object",
                        "example": {
                            "power": "24kW"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
}